+++ packages/website/src/_services/api/activity-overview-service/types.ts
@@ -1,4 +1,5 @@
 import { DatasetDefinition } from '~/api'
+import { TimeBucket } from '~/api/species-richness-service'
 import { ApiHourlySpeciesSummary } from '~/api-helpers/mock'
 
 export interface DetectionGroupedBySiteAndTaxon {
@@ -13,6 +14,8 @@ export interface DetectionGroupByDetectionKey {
 
 export interface ActivityOverviewData extends DatasetDefinition {
   overviewBySite: ActicvityOverviewDataBySite
+  overviewByTime: ActivityOverviewDataByTime[]
+  overviewBySpecies: ActivityOverviewDataBySpecies[]
 }
 
 export interface ActicvityOverviewDataBySite {
@@ -27,4 +30,27 @@ export interface ActicvityOverviewDataBySite {
       occupancy: boolean
     }
   }
+}
+
+export type ActivityOverviewDataByTime = Record<TimeBucket, ActivityOverviewDataByTimeBucket>
+
+export interface ActivityOverviewDataByTimeBucket {
+  detection: Record<number, number>
+  detectionFrequency: Record<number, number>
+  occupancy: Record<number, number>
+}
+
+export const ACTIVITY_OVERVIEW_TIME_KEYS: Record<string, keyof ActivityOverviewDataByTimeBucket> = {
+  detection: 'detection',
+  detectionFrequency: 'detectionFrequency',
+  occupancy: 'occupancy'
+}
+
+export interface ActivityOverviewDataBySpecies {
+  speciesName: string
+  taxonomyClass: string
+  detectionCount: number
+  detectionFrequency: number
+  occupiedSites: number
+  occupancyNaive: number
 }