+++ packages/website/src/_services/dataset-filters/comparison-list/comparison-list.ts
@@ -1,22 +1,24 @@
 import { Options, Vue } from 'vue-class-component'
-import { Emit, Inject } from 'vue-property-decorator'
+import { Emit, Inject, Prop } from 'vue-property-decorator'
 
 import { dayjs } from '@rfcx-bio/utils/dayjs-initialized'
 
 import { BiodiversityStore } from '~/store'
-import { ColoredFilter, Filter } from '..'
+import { ColoredFilter, ComparisonFilter } from '..'
 import { FilterImpl } from '../classes'
 import ComparisonFilterModalComponent from '../comparison-filter-modal/comparison-filter-modal.vue'
 
-const defaultFilter = new FilterImpl(dayjs().subtract(20, 'years'), dayjs(), [])
+const defaultFilter = new FilterImpl(dayjs().subtract(20, 'years'), dayjs())
 
 @Options({
   components: {
     ComparisonFilterModalComponent
   }
 })
+
 export default class ComparisonListComponent extends Vue {
   @Inject() readonly store!: BiodiversityStore
+  @Prop({ default: true }) canFilterByTaxon!: boolean
   @Emit() emitSelect (): ColoredFilter[] {
     return this.filters.map((f, i) => ({
       ...f,
@@ -28,7 +30,7 @@ export default class ComparisonListComponent extends Vue {
   isAddSelected = false
   isFilterOpen = false
   filters: FilterImpl[] = [defaultFilter]
-  currentSelectedFilter: FilterImpl | null = null
+  modalFilter: FilterImpl | null = null
 
   //  TODO: Have to improve this logic to check what is `all` meaning
   get isDefaultFilter (): boolean {
@@ -46,10 +48,11 @@ export default class ComparisonListComponent extends Vue {
   addFilterConfig (): void {
     // Copy previous filter
     const previousFilter = this.filters[this.filters.length - 1]
-    this.currentSelectedFilter = new FilterImpl(
+    this.modalFilter = new FilterImpl(
       previousFilter.startDate,
       previousFilter.endDate,
-      [...previousFilter.sites],
+      previousFilter.sites.map(s => ({ ...s })),
+      previousFilter.otherFilters.map(f => ({ ...f })),
       previousFilter.color
     )
 
@@ -62,11 +65,27 @@ export default class ComparisonListComponent extends Vue {
     return this.store.datasetColors[idx]
   }
 
+  getOptionalFilterText (idx: number): string {
+    const otherFilters = this.filters[idx].otherFilters
+    if (otherFilters.length === 1) {
+      return `${otherFilters[0].propertyName}: ${otherFilters[0].value}`
+    } else {
+      return `+ ${otherFilters.length} filter${otherFilters.length > 1 ? 's' : ''} applied`
+    }
+    // TODO: 268 Show full information of filter when the user hovers over the comparison box
+    /*
+    const optionalFilters = groupBy(this.filters[idx].otherFilters, 'title')
+    const getFilterValue = (filters: OptionalFilter[]): string => filters.map(f => f.value).join(', ')
+    const groupedOptionalFilters = mapValues(optionalFilters, getFilterValue) // type ==> { [x: string]: string } ==> e.g. {'taxon': 'frog, bird'}
+    return groupedOptionalFilters // TODO: expect to transform this { [x: string]: string } to string ==> e.g. "Taxon: frog, bird"
+    */
+  }
+
   popupOpen (idx: number): void {
     this.isFilterOpen = true
     this.isAddSelected = false
     this.selectedFilterId = idx
-    this.currentSelectedFilter = this.filters[this.selectedFilterId]
+    this.modalFilter = this.filters[this.selectedFilterId]
   }
 
   popupClose (): void {
@@ -81,8 +100,8 @@ export default class ComparisonListComponent extends Vue {
     this.emitSelect()
   }
 
-  apply (filter: Filter): void {
-    const newFilter = new FilterImpl(filter.startDate, filter.endDate, filter.sites)
+  apply (filter: ComparisonFilter): void {
+    const newFilter = new FilterImpl(filter.startDate, filter.endDate, filter.sites, filter.otherFilters)
     if (this.isAddSelected) {
       this.filters.push(newFilter)
       this.isAddSelected = false