+++ packages/website/src/_services/maps/map-tool-menu/map-tool-menu.ts
@@ -1,21 +1,15 @@
 import { Vue } from 'vue-class-component'
-import { Emit, Prop, Watch } from 'vue-property-decorator'
+import { Emit, Prop } from 'vue-property-decorator'
 
-import { MAPBOX_STYLE_WITH_PLACE_LABELS } from '~/maps'
-import { TAXONOMY_CLASSES } from '../../api/taxonomy-service'
+import { MAPBOX_STYLE_RFCX_WITH_PLACE_LABELS, MAPBOX_STYLE_SATELLITE_STREETS, MapboxStyle } from '~/maps'
 
 interface MapOptions {
-  id: string
+  style: MapboxStyle
   name: string
 }
 
 export default class MapToolMenuComponent extends Vue {
-  @Prop({ default: false }) displayTaxonomies!: boolean
-
-  @Emit()
-  emitTaxonomyValue (): string {
-    return this.taxon
-  }
+  @Prop() mapStyle!: MapboxStyle
 
   @Emit()
   emitShowLabelsToggle (): boolean {
@@ -24,33 +18,13 @@ export default class MapToolMenuComponent extends Vue {
   }
 
   @Emit()
-  emitMapStyle (): string {
-    return this.mapStyle
+  emitMapStyle (style: MapboxStyle): MapboxStyle {
+    return style
   }
 
   isShowLabels = true
-  mapStyle = 'mapbox://styles/mapbox/satellite-streets-v11'
-  taxons = TAXONOMY_CLASSES
-  taxon = this.taxons[0].name
-
-  get mapOptions (): MapOptions[] {
-    return [
-      { id: 'mapbox://styles/mapbox/satellite-streets-v11', name: 'Satellite' },
-      { id: MAPBOX_STYLE_WITH_PLACE_LABELS, name: 'Simple' }
-    ]
-  }
-
-  @Watch('taxon')
-  onTaxononyChange (): void {
-    this.emitTaxonomyValue()
-  }
-
-  @Watch('mapStyle')
-  onMapStyleChange (): void {
-    this.emitMapStyle()
-  }
-
-  setMapStyle (id: string): void {
-    this.mapStyle = id
-  }
+  mapStyleOptions: MapOptions[] = [
+    { style: MAPBOX_STYLE_SATELLITE_STREETS, name: 'Satellite' },
+    { style: MAPBOX_STYLE_RFCX_WITH_PLACE_LABELS, name: 'Simple' }
+  ]
 }