+++ packages/website/src/_services/vue/vite-env.d.ts
@@ -3,7 +3,5 @@ import 'vite/client'
 declare global {
   interface ImportMetaEnv extends Readonly<Record<string, string>> {
     readonly VITE_BIO_API_HOST: string
-
-    readonly VITE_TOGGLE_AO_LINE: string
   }
 }