+++ CHANGELOG.md
@@ -12,7 +12,10 @@ Allowed categories:
 
 # 0.0.9 (2021-11-25)
 
-To be update
+- **activity-overview:** Group detection, detection frequency, and occupancy map [#255][#257]
+- **activity-overview:** Group detection, detection frequency, and occupancy map [#259][#260]
+- **selection:** Add taxon in dataset comparison filter [#219]
+- **selection:** Remove taxon toggle from map [#267]
 
 # 0.0.8 (2021-11-22)